isWasm
